+++ src/document/query/any/index.ts
@@ -1,0 +1,1 @@
+export { AnyQuery } from './any-query'