+++ src/document/query/index.ts
@@ -1,3 +1,4 @@
 export { Query } from './main'
+export { AnyQuery } from './any'
 export * from './get'
 export * from './find'