+++ src/document/transaction/any/index.ts
@@ -1,0 +1,2 @@
+export { AnyTransaction } from './any-transaction'
+export type { AnyAddChildParams, AnyUpdateParams } from './any.types'