+++ src/document/transaction/index.ts
@@ -1,6 +1,8 @@
 export { Transaction } from './main'
+export { AnyTransaction } from './any'
 export { isTransactionContext } from './guard'
 
+export type * from './any'
 export type * from './clone'
 export type * from './create'
 export type * from './update'